requirements.txt
PyQt5==5.15.9
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.0.0
lxml==4.9.3 